requirements.txt
Django==3.0.6
djangorestframework==3.13.1
plaid==0.1.7
plaid_python==4.0.0
